Fixing error on Arit

+++ src/robot/Arit.java
@@ -71,13 +71,13 @@ final public class Arit
         {
             Coordinate a = (Coordinate) arg1, b = (Coordinate) arg2;
             
-            double a1 = a.getCoordinate()[0];
-            double a2 = a.getCoordinate()[1];
-            double b1 = b.getCoordinate()[0];
-            double b2 = b.getCoordinate()[1];
+            int a1 = a.getCoordinate()[0];
+            int a2 = a.getCoordinate()[1];
+            int b1 = b.getCoordinate()[0];
+            int b2 = b.getCoordinate()[1];
             
-            double ans1 = op.op(b1, a1);
-            double ans2 = op.op(b2, a2);
+            int ans1 = (int) op.op(b1, a1);
+            int ans2 = (int) op.op(b2, a2);
             rvm.DATA.push(new Coordinate(ans1, ans2));
         }
         else { throw new WrongTypeException("Num or Coordinate"); }
